feat(DTFS2-6734): convert admin/activity route to use typescript
