+++ portal/server/routes/admin/index.ts
@@ -1,0 +1,16 @@
+import express from 'express';
+
+import userAdminRoutes from './users';
+// eslint-disable-next-line import/no-unresolved, import/extensions
+import activityRoutes from './activity';
+
+import { validateToken, validateRole } from '../middleware';
+import { ADMIN } from '../../constants/roles';
+
+const router = express.Router();
+
+router.use('/admin/*', [validateToken, validateRole({ role: [ADMIN] })]);
+router.use('/admin/', userAdminRoutes);
+router.use('/admin/', activityRoutes);
+
+export default router;
