+++ portal/server/routes/index.ts
@@ -1,0 +1,32 @@
+import express from 'express';
+import startRoutes from './start';
+import loginRoutes from './login';
+import dashboardRoutes from './dashboard';
+import contractRoutes from './contract';
+import eligibilityRoutes from './contract/eligibility';
+import footerRoutes from './footer.route';
+// eslint-disable-next-line import/extensions, import/no-unresolved
+import adminRoutes from './admin';
+import userRoutes from './user';
+import feedbackRoutes from './feedback';
+import schemeTypeRoutes from './schemeType';
+import portalRoutes from './reports.route';
+import staticRoutes from './static';
+
+const router = express.Router();
+
+router.use('/', startRoutes);
+router.use('/', loginRoutes);
+router.use('/', dashboardRoutes);
+router.use('/', contractRoutes);
+router.use('/', eligibilityRoutes);
+router.use('/', footerRoutes);
+router.use('/', adminRoutes);
+router.use('/', userRoutes);
+router.use('/', feedbackRoutes);
+router.use('/', schemeTypeRoutes);
+router.use('/', portalRoutes);
+router.use('/', feedbackRoutes);
+router.use('/', staticRoutes);
+
+export default router;
